Refactor for Socrata Source Data Starting to refactor the API to test feasability of querying Socrata (City of Austin Open Data Portal) for trip records.
Under this model, we replace the part of the app logic which aggregates trip counts from in-memory json data with a routine that fetches grouped trip counts from Socrata. Socrata hosts the individual trip records and supports SQL-like querying, so this opens up the possibiltiy of querying by datetime (or any other trip property). The app will translate the Socrata query results into a geojson grid + trip count response, returning a payload in the same format it returns today. In other words, this can be done without a breaking change to the API.

The primary technical concern is that the app will now be making an HTTP request behind the scenes that will definitely add latency—hopefuly just a second or two. We'll see.